implementando integração com raspay

+++ src/main/java/com/client/ws/rasmooplus/dto/wsraspay/OrderDto.java
@@ -4,6 +4,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
@@ -11,7 +13,7 @@ public class OrderDto {
 
     private String id;
     private String customerId;
-    private Long discount;
+    private BigDecimal discount;
     private String productAcronym;
 
 }
